+++ extensions/RaiseSuperHeroAgent/main.ts
@@ -65,4 +65,32 @@ namespace superHeroAgent {
         player.say("Mission complete!")
         player.say("SuperHero team success!")
     }
+
+    //% blockId=superhero_fight
+    //% block="SuperHero Agent fight"
+    //% weight=86
+    export function fight(): void {
+        player.say("#sphr fight")
+    }
+
+    //% blockId=superhero_heal_player
+    //% block="SuperHero Agent heal player"
+    //% weight=84
+    export function healPlayer(): void {
+        player.say("#sphr heal")
+    }
+
+    //% blockId=superhero_defend
+    //% block="SuperHero Agent defend"
+    //% weight=82
+    export function defend(): void {
+        player.say("#sphr defend")
+    }
+
+    //% blockId=superhero_light
+    //% block="SuperHero Agent light mode"
+    //% weight=80
+    export function lightMode(): void {
+        player.say("#sphr light")
+    }
 }